Add Python 3.7 to .travis.yml and setup.py
This commit adds Python 3.7 references to the .travis.yml and setup.py
files. The commit also replaces a reference in the .travis.yml for
installing the dxlfiletransferclient from a git branch. This is no longer
needed since the dxlfiletransferclient package has now been released to
PyPI. The commit additionally includes some minor updates to the
.gitignore file to exclude some files used for client provisioning.